+++ exporter/exporter_suite_test.go
@@ -1,4 +1,4 @@
-package app_test
+package exporter_test
 
 import (
 	"testing"
@@ -7,7 +7,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestApp(t *testing.T) {
+func TestExporter(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "App Suite")
+	RunSpecs(t, "Exporter Suite")
 }
